+++ jac/jaclang/compiler/program.py
@@ -15,11 +15,9 @@ from jaclang.compiler.passes.main import (
     CompilerMode,
     DeclImplMatchPass,
     DefUsePass,
-    FuseTypeInfoPass,
     InheritancePass,
     JacAnnexPass,
     JacImportPass,
-    JacTypeCheckPass,
     PyBytecodeGenPass,
     PyCollectDepsPass,
     PyImportPass,
@@ -151,13 +149,6 @@ class JacProgram:
         for mod in self.mod.hub.values():
             self.schedule_runner(mod, mode=CompilerMode.COMPILE)
 
-        # Check if we need to run without type checking then just return
-        if mode == CompilerMode.COMPILE:
-            return mod_targ
-
-        # Run TypeCheckingPass on the top module
-        JacTypeCheckPass(mod_targ, prog=self)
-
         for mod in self.mod.hub.values():
             PyCollectDepsPass(mod, prog=self)
 
@@ -174,6 +165,10 @@ class JacProgram:
 
         for mod in self.mod.hub.values():
             DefUsePass(mod, prog=self)
+
+        # Check if we need to run without type checking then just return
+        if mode == CompilerMode.COMPILE:
+            return mod_targ
 
         for mod in self.mod.hub.values():
             self.schedule_runner(mod, mode=CompilerMode.TYPECHECK)
@@ -194,9 +189,10 @@ class JacProgram:
         py_code_gen = [
             PyastGenPass,
             PyJacAstLinkPass,
+            InheritancePass,
+            AccessCheckPass,
             PyBytecodeGenPass,
         ]
-        type_checker_sched = [InheritancePass, FuseTypeInfoPass, AccessCheckPass]
 
         final_pass: Optional[type[Transform[uni.Module, uni.Module]]] = None
         match mode:
@@ -204,8 +200,6 @@ class JacProgram:
                 passes = ir_gen_sched
             case CompilerMode.COMPILE:
                 passes = [*ir_gen_sched, *py_code_gen]
-            case CompilerMode.TYPECHECK:
-                passes = type_checker_sched
             case _:
                 raise ValueError(f"Invalid mode: {mode}")
         for current_pass in passes:
