+++ jac/jaclang/tests/test_language.py
@@ -105,39 +105,6 @@ class JacLanguageTests(TestCase):
             "Too high!\nToo low!\nToo high!\nCongratulations! You guessed correctly.\n",
         )
 
-    def test_multi_dim_arr_slice(self) -> None:
-        """Parse micro jac file."""
-        captured_output = io.StringIO()
-        sys.stdout = captured_output
-        cli.tool(
-            "ir",
-            [
-                "ast",
-                self.fixture_abs_path("multi_dim_array_split.jac"),
-            ],
-        )
-        sys.stdout = sys.__stdout__
-        stdout_value = captured_output.getvalue()
-
-        expected_outputs = [
-            "+-- AtomTrailer - Type: builtins.list[builtins.int]",
-            "    +-- Name - arr - Type: builtins.list[builtins.list[builtins.int]],  SymbolTable: list",
-            "+-- IndexSlice - [IndexSlice] - Type: builtins.list[builtins.list[builtins.int]],  SymbolTable: None",
-            "        +-- Token - [, ",
-            "        +-- Int - 1 - Type: Literal[1]?,  SymbolTable: None",
-            "        +-- Token - :, ",
-            "        +-- Int - 3 - Type: Literal[3]?,  SymbolTable: None",
-            "        +-- Token - ,, ",
-            "        +-- Int - 1 - Type: Literal[1]?,  SymbolTable: None",
-            "        +-- Token - :, ",
-            "        +-- Token - :, ",
-            "        +-- Int - 2 - Type: Literal[2]?,  SymbolTable: None",
-            "        +-- Token - ], ",
-        ]
-
-        for expected in expected_outputs:
-            self.assertIn(expected, stdout_value)
-
     def test_dotgen(self) -> None:
         """Test the dot gen of builtin function."""
         import json
@@ -330,7 +297,7 @@ class JacLanguageTests(TestCase):
             mach, "deep_import_interp", base_path=self.fixture_abs_path("./")
         )
         print(mach.jac_program.mod.hub.keys())
-        self.assertEqual(len(mach.jac_program.mod.hub.keys()), 5)
+        self.assertEqual(len(mach.jac_program.mod.hub.keys()), 6)
 
     def test_deep_imports_mods(self) -> None:
         """Parse micro jac file."""
@@ -550,49 +517,6 @@ class JacLanguageTests(TestCase):
         stdout_value = captured_output.getvalue()
         self.assertIn("2 Accessing privileged Data", stdout_value)
 
-    def test_needs_import_1(self) -> None:
-        """Test py ast to Jac ast conversion output."""
-        file_name = self.fixture_abs_path("pyfunc_1.py")
-
-        from jaclang.compiler.passes.main.pyast_load_pass import PyastBuildPass
-        import ast as py_ast
-        import jaclang.compiler.unitree as uni
-
-        with open(file_name, "r") as f:
-            file_source = f.read()
-            parsed_ast = py_ast.parse(file_source)
-            try:
-                py_ast_build_pass = PyastBuildPass(
-                    ir_in=uni.PythonModuleAst(
-                        parsed_ast, orig_src=uni.Source(file_source, file_name)
-                    ),
-                    prog=JacProgram(),
-                ).ir_out
-            except Exception as e:
-                return f"Error While Jac to Py AST conversion: {e}"
-
-        (prog := JacProgram()).compile_from_str(
-            source_str=py_ast_build_pass.unparse(),
-            file_path=file_name[:-3] + ".jac",
-            mode=CMode.TYPECHECK,
-        )
-
-        architype_count = 0
-        for mod in prog.mod.hub.values():
-            if mod.name == "builtins":
-                continue
-            architype_count += len(mod.get_all_sub_nodes(uni.Architype))
-
-        self.assertEqual(architype_count, 24)
-        captured_output = io.StringIO()
-        sys.stdout = captured_output
-        Jac.jac_import(
-            self.mach, "needs_import_1", base_path=self.fixture_abs_path("./")
-        )
-        sys.stdout = sys.__stdout__
-        stdout_value = captured_output.getvalue()
-        self.assertIn("pyfunc_1 imported", stdout_value)
-
     def test_pyfunc_1(self) -> None:
         """Test py ast to Jac ast conversion."""
         from jaclang.compiler.passes.main import PyastBuildPass
@@ -619,51 +543,6 @@ class JacLanguageTests(TestCase):
             '\n\n@ my_decorator \n def say_hello() {\n\n    """Say hello""" ;', output
         )
 
-    def test_needs_import_2(self) -> None:
-        """Test py ast to Jac ast conversion output."""
-        file_name = self.fixture_abs_path("pyfunc_2.py")
-
-        from jaclang.compiler.passes.main.pyast_load_pass import PyastBuildPass
-        import ast as py_ast
-        import jaclang.compiler.unitree as uni
-
-        with open(file_name, "r") as f:
-            file_source = f.read()
-            parsed_ast = py_ast.parse(file_source)
-            try:
-                py_ast_build_pass = PyastBuildPass(
-                    ir_in=uni.PythonModuleAst(
-                        parsed_ast,
-                        orig_src=uni.Source(file_source, file_name),
-                    ),
-                    prog=JacProgram(),
-                ).ir_out
-            except Exception as e:
-                return f"Error While Jac to Py AST conversion: {e}"
-
-            (prog := JacProgram()).compile_from_str(
-                source_str=py_ast_build_pass.unparse(),
-                file_path=file_name[:-3] + ".jac",
-                mode=CMode.TYPECHECK,
-            )
-
-        architype_count = 0
-        for mod in prog.mod.hub.values():
-            if mod.name == "builtins":
-                continue
-            architype_count += len(mod.get_all_sub_nodes(uni.Architype))
-
-        self.assertEqual(architype_count, 30)  # Because of the Architype from math
-        captured_output = io.StringIO()
-        sys.stdout = captured_output
-        Jac.jac_import(
-            self.mach, "needs_import_2", base_path=self.fixture_abs_path("./")
-        )
-        sys.stdout = sys.__stdout__
-        stdout_value = captured_output.getvalue()
-        self.assertIn("pyfunc_2 imported", stdout_value)
-        self.assertEqual(stdout_value.count("<class 'bytes'>"), 3)
-
     def test_pyfunc_2(self) -> None:
         """Test py ast to Jac ast conversion."""
         from jaclang.compiler.passes.main import PyastBuildPass
@@ -684,41 +563,6 @@ class JacLanguageTests(TestCase):
         self.assertIn("br = b'Hello\\\\\\\\nWorld'", output)
         self.assertIn("class Circle {\n    def init(radius: float", output)
         self.assertIn("<>node = 90;    \n    print(<>node) ;\n}\n", output)
-
-    def test_needs_import_3(
-        self,
-    ) -> None:  # TODO : Pyfunc_3 has a bug in conversion in matchmapping node
-        """Test py ast to Jac ast conversion output."""
-        file_name = self.fixture_abs_path("pyfunc_3.py")
-        import jaclang.compiler.unitree as uni
-
-        with open(file_name, "r") as f:
-            file_source = f.read()
-        (prog := JacProgram()).compile_from_str(
-            source_str=file_source, file_path=file_name, mode=CMode.TYPECHECK
-        )
-
-        architype_count = sum(
-            len(mod.get_all_sub_nodes(uni.Architype))
-            for mod in prog.mod.hub.values()
-            if mod.name != "builtins"
-        )
-        self.assertEqual(
-            architype_count, 58
-        )  # Fixed duplication of 'case' module (previously included 3 times, added 20 extra Architypes; 75 → 55)
-        builtin_mod = next(
-            (mod for name, mod in prog.mod.hub.items() if "builtins" in name),
-            None,
-        )
-        self.assertEqual(len(builtin_mod.get_all_sub_nodes(uni.Architype)), 109)
-        captured_output = io.StringIO()
-        sys.stdout = captured_output
-        Jac.jac_import(
-            self.mach, "needs_import_3", base_path=self.fixture_abs_path("./")
-        )
-        sys.stdout = sys.__stdout__
-        stdout_value = captured_output.getvalue()
-        self.assertIn("pyfunc_3 imported", stdout_value)
 
     def test_pyfunc_3(self) -> None:
         """Test py ast to Jac ast conversion."""
@@ -788,29 +632,6 @@ class JacLanguageTests(TestCase):
         prog.format()
         after = prog.format()
         self.assertEqual(before, after)
-
-    def test_type_fuse_expr(self) -> None:
-        """Basic test for pass."""
-        captured_output = io.StringIO()
-        sys.stdout = captured_output
-        cli.tool(
-            "ir",
-            [
-                "ast",
-                self.examples_abs_path("reference/collection_values.jac"),
-            ],
-        )
-
-        sys.stdout = sys.__stdout__
-        stdout_value = captured_output.getvalue()
-        self.assertIn(
-            "builtins.dict[builtins.int, builtins.int]",
-            stdout_value,
-        )
-        self.assertIn(
-            "typing.Generator[builtins.int, None, None]",
-            stdout_value,
-        )
 
     def test_inherit_check(self) -> None:
         """Test py ast to Jac ast conversion output."""
@@ -882,59 +703,6 @@ class JacLanguageTests(TestCase):
         self.assertEqual("9", stdout_value[0])
         self.assertEqual("567", stdout_value[1])
 
-    def test_random_check(self) -> None:
-        """Test py ast to Jac ast conversion output."""
-        from jaclang.settings import settings
-
-        module_paths = ["random", "ast"]
-        for module_path in module_paths:
-            stdlib_dir = sysconfig.get_paths()["stdlib"]
-            file_path = os.path.join(
-                stdlib_dir,
-                module_path + ".py",
-            )
-            settings.print_py_raised_ast = True
-            with open(file_path) as f:
-                file_source = f.read()
-            ir = JacProgram().compile_from_str(
-                source_str=file_source,
-                file_path=file_path,
-                mode=CMode.TYPECHECK,
-            )
-            gen_ast = ir.pp()
-            if module_path == "random":
-                self.assertIn("ModulePath - statistics -", gen_ast)
-            else:
-                self.assertIn("+-- Name - NodeTransformer - Type: No", gen_ast)
-
-    def test_deep_convert(self) -> None:
-        """Test py ast to Jac ast conversion output."""
-        file_name = self.fixture_abs_path("pyfunc_1.py")
-
-        import jaclang.compiler.unitree as uni
-        from jaclang.settings import settings
-
-        settings.print_py_raised_ast = True
-        with open(file_name, "r") as f:
-            file_source = f.read()
-        ir = (prog := JacProgram()).compile_from_str(
-            source_str=file_source, file_path=file_name, mode=CMode.TYPECHECK
-        )
-        jac_ast = ir.pp()
-        self.assertIn(" |   +-- String - 'Loop completed normally{}'", jac_ast)
-        sub_node_list_count = 0
-        for i in prog.mod.hub.values():
-            if i.name == "builtins":
-                continue
-            sub_node_list_count += len(i.get_all_sub_nodes(uni.SubNodeList))
-        self.assertEqual(sub_node_list_count, 623)
-        captured_output = io.StringIO()
-        sys.stdout = captured_output
-        Jac.jac_import(self.mach, "deep_convert", base_path=self.fixture_abs_path("./"))
-        sys.stdout = sys.__stdout__
-        stdout_value = captured_output.getvalue()
-        self.assertIn("Deep convo is imported", stdout_value)
-
     def test_override_walker_inherit(self) -> None:
         """Test py ast to Jac ast conversion output."""
         captured_output = io.StringIO()
@@ -945,34 +713,6 @@ class JacLanguageTests(TestCase):
         sys.stdout = sys.__stdout__
         stdout_value = captured_output.getvalue()
         self.assertEqual("baz\nbar\n", stdout_value)
-
-    def test_ds_type_check_pass(self) -> None:
-        """Test conn assign on edges."""
-        (mypass := JacProgram()).compile(
-            self.examples_abs_path("micro/simple_walk.jac"),
-            mode=CMode.TYPECHECK,
-        )
-        self.assertEqual(len(mypass.errors_had), 0)
-        self.assertEqual(len(mypass.warnings_had), 0)
-
-    def test_ds_type_check_pass2(self) -> None:
-        """Test conn assign on edges."""
-        (mypass := JacProgram()).compile(
-            self.examples_abs_path("guess_game/guess_game5.jac"),
-            mode=CMode.TYPECHECK,
-        )
-        self.assertEqual(len(mypass.errors_had), 0)
-        self.assertEqual(len(mypass.warnings_had), 0)
-
-    def test_circle_override1_type_check_pass(self) -> None:
-        """Test conn assign on edges."""
-        (mypass := JacProgram()).compile(
-            self.examples_abs_path("manual_code/circle.jac"),
-            mode=CMode.TYPECHECK,
-        )
-        self.assertEqual(len(mypass.errors_had), 0)
-        # FIXME: Figure out what to do with warning.
-        # self.assertEqual(len(mypass.warnings_had), 0)
 
     def test_self_with_no_sig(self) -> None:  # we can get rid of this, isn't?
         """Test py ast to Jac ast conversion output."""
